requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ML / Deep Learning
torch==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==1.5.3
shap==0.44.1

# AI Assistant
transformers==4.37.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# CORS
starlette==0.35.1
